test/requirements.txt
anyio==4.5.0
asyncio==3.4.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.7
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
PySocks==1.7.1
pytest==8.3.2
pytest-json-report==1.5.0
pytest-metadata==3.1.1
requests==2.32.3
selenium==4.24.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
websockets==13.0
wsproto==1.2.0
